refactor: move functions from component to utils
